Fix a couple of compiler warnings related to header_rewrite and traffic_cop changes

+++ cmd/traffic_top/stats.h
@@ -67,8 +67,8 @@ public:
       }
 
       // set the host
-      int start = _url.find(":");
-      int end = _url.find("/", start + 3);
+      size_t start = _url.find(":");
+      size_t end = _url.find("/", start + 3);
       _host = _url.substr(start + 3, end - start - 3);
       end = _host.find(":");
       if (end != string::npos) {
